requirements.txt
datasets==2.19.0
einops==0.8.1
evaluate==0.4.2
flash_attn==2.6.1
fsspec==2024.2.0
huggingface_hub==0.34.4
hydra-core==1.3.2
lightning==2.5.5
numpy==1.24.3
omegaconf==2.3.0
Requests==2.32.5
rich==14.2.0
scipy==1.16.2
tokenizers==0.22.1
torch==2.3.0
torchmetrics==1.8.2
tqdm==4.66.4
transformers==4.56.1
